% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importXlsxFolder.R
\name{importXlsxSep}
\alias{importXlsxSep}
\title{Import all excel files of specific folders to a list of data frames}
\usage{
importXlsxSep(dataPath, rowsToImport = "all", colsToImport = "all")
}
\arguments{
\item{dataPath}{Location of the folder containing the data}

\item{rowsToImport}{First row to import (default = "all").}

\item{colsToImport}{First column to import (default = "all").}
}
\value{
Return a list of dataframes containing the informations from the first row and column to the end of the files.
}
\description{
Import all excel files of specific folders to a list of data frames
}
\examples{
importXlsxSep("./datas/myFolder/")

}
